Rename BackendFramework.ValueModels namespace to match C# guidelines
