+++ Backend.Tests/LiftControllerTests.cs
@@ -11,6 +11,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using BackendFramework.Models;
 
 namespace Backend.Tests
 {
