+++ Backend.Tests/UserRoleControllerTests.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
+using BackendFramework.Models;
 
 namespace Backend.Tests
 {
