+++ Backend.Tests/WordControllerTests.cs
@@ -7,6 +7,7 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using BackendFramework.Models;
 
 namespace Backend.Tests
 {
