+++ Backend/Controllers/UserEditController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Controllers
 {
