+++ Backend/Controllers/UserRoleController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Controllers
 {
