+++ Backend/Controllers/WordController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Controllers
 {
