+++ Backend/Interfaces/ISemDomParser.cs
@@ -1,6 +1,7 @@
 ﻿using BackendFramework.ValueModels;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Interfaces
 {
