+++ Backend/Models/Project.cs
@@ -1,12 +1,13 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using BackendFramework.ValueModels;
+using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace BackendFramework.ValueModels
+namespace BackendFramework.Models
 {
     public class Project
     {
