+++ Backend/Services/SemDomParser.cs
@@ -3,6 +3,7 @@ using BackendFramework.ValueModels;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Services
 {
