+++ Backend/Services/WordRepository.cs
@@ -3,6 +3,7 @@ using BackendFramework.ValueModels;
 using MongoDB.Driver;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using BackendFramework.Models;
 
 namespace BackendFramework.Services
 {
